pyproject.toml
[project]
name = "medical-agent-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "assemblyai>=0.46.0",
    "fastapi>=0.121.0",
    "livekit-agents[audio,core]",
    "livekit-api>=1.0.7",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "llama-index>=0.14.7",
    "logfire>=4.14.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.11.1",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
